chore(CHANGELOG): add changelogs for v2.0.0 (#35)
Along with Changelog updates, this commit fixes the travis
 - bumps the k8s version in minikube
 - updates the latest cstor operator
   yaml with new CRDs
 - improves error reporting by adding logs to tests

+++ ci/migrate/sanity.sh
@@ -7,6 +7,11 @@ make migrate-image.amd64
 # setup openebs & cstor v1 for migration 
 ./ci/migrate/setup.sh || exit 1
 # run migration tests
-./ci/migrate/test.sh || exit 1
+./ci/migrate/test.sh 
+if [[ $? != 0 ]]; then
+  kubectl logs --tail=50 -l job-name=migrate-pool -n openebs
+  kubectl logs --tail=50 -l job-name=migrate-volume -n openebs
+  exit 1
+fi
 
 rm ./ci/migrate/volume.yaml ./ci/migrate/application.yaml
